fix compile error

+++ be/src/vec/functions/function_string.h
@@ -1798,7 +1798,7 @@ public:
     size_t get_number_of_arguments() const override { return 2; }
 
     Status execute_impl(FunctionContext* context, Block& block, const ColumnNumbers& arguments,
-                        uint32_t result, size_t input_rows_count) const override {
+                        size_t result, size_t input_rows_count) const override {
         auto res_column = ColumnString::create();
         ColumnPtr argument_column = block.get_by_position(arguments[0]).column;
         ColumnPtr argument_column_2;
